feat(#16): widen __heap_bump_bytes() from i32 to i64 (#6027)
Every backend's runtime already computed the bump high-water mark in 64
bits — the natives subtract in rax / x0, and wasm's reader had the full
i32 address range. Only the *declared* result was i32, and that was
enough to truncate: a quadratic sweep read 141 MB / 555 MB / -2.09 GB /
202 MB, where only the third of the four readings looks wrong. The arena
is 16 GiB, so the mark passes 2^31 on exactly the long runs the probe
exists to measure.

checker.FuncSigs and the self-host's asmcore.infer_expr_type both say i64
now; the self-host previously fell through to `unknown`, which let the two
compilers disagree unnoticed. The natives needed no codegen change, but
wasm did: both operands are i32 linear-memory addresses, so the helper
subtracts in i32 and appends i64.extend_i32_u — zero-, not sign-extend.
irlower's infer_expr_width and lower_i64 learn the builtin so the
`as i32` narrowing every rc probe uses routes through int_wrap.

The ~540 existing probe sites gained an explicit `as i32`, preserving each
assertion exactly; arm64_asm_bench_run.fern keeps the full i64 via std/i64
rather than reintroducing the bug where it actually bites.

Gates: rc_heap_bump_i64_test.go — a type-level leg on all three backends,
plus a 2400000032-byte reading that came out as -1894967264 before.

+++ internal/e2e/append_alias_differential_test.go
@@ -163,10 +163,10 @@ function main(): i32 {
     var acc: i32 = 0;
     var w: i32 = 0;
     while (w < 200) { acc = acc + take(path.append(w)); w = w + 1; }
-    var b1: i32 = __heap_bump_bytes();
+    var b1: i32 = (__heap_bump_bytes() as i32);
     var i: i32 = 0;
     while (i < 5000) { acc = acc + take(path.append(i)); i = i + 1; }
-    var b2: i32 = __heap_bump_bytes();
+    var b2: i32 = (__heap_bump_bytes() as i32);
     if (__rc_underflow_count() != 0) { return 99; }
     if (b2 - b1 >= 512) { return 98; }
     if (path.len() != 2) { return 97; }
