+++ internal/e2e/rc_heap_bump_dyn_trait_x86_64_test.go
@@ -35,7 +35,7 @@ impl Shape for Boxed {
     function area(self: Self): i32 { return 1; }
 }
 function main(): i32 {
-    var before: i32 = __heap_bump_bytes();
+    var before: i32 = (__heap_bump_bytes() as i32);
     var i: i32 = 0;
     var sum: i32 = 0;
     while (i < ` + n + `) {
@@ -43,7 +43,7 @@ function main(): i32 {
         sum = sum + d.area();
         i = i + 1;
     }
-    return __heap_bump_bytes() - before;
+    return (__heap_bump_bytes() as i32) - before;
 }`
 }
 
@@ -107,7 +107,7 @@ impl B for Both {
     function b2(self: Self): i32 { return 3; }
 }
 function main(): i32 {
-    var before: i32 = __heap_bump_bytes();
+    var before: i32 = (__heap_bump_bytes() as i32);
     var i: i32 = 0;
     var sum: i32 = 0;
     while (i < ` + n + `) {
@@ -115,7 +115,7 @@ function main(): i32 {
         sum = sum + d.a1() + d.b1() + d.b2();
         i = i + 1;
     }
-    return __heap_bump_bytes() - before;
+    return (__heap_bump_bytes() as i32) - before;
 }`
 	}
 	small := mustRunX86_64FreeOn(t, src("50"))
@@ -166,7 +166,7 @@ function main(): i32 {
 	// Bounded check: the caller's local still reclaims; no leak from the
 	// borrow (and no double-free corrupting the freelist).
 	bumped := func(n string) string {
-		return mk(`var before: i32 = __heap_bump_bytes();
+		return mk(`var before: i32 = (__heap_bump_bytes() as i32);
     var i: i32 = 0;
     var sum: i32 = 0;
     while (i < ` + n + `) {
@@ -174,7 +174,7 @@ function main(): i32 {
         sum = sum + use_it(d);
         i = i + 1;
     }
-    return __heap_bump_bytes() - before;`)
+    return (__heap_bump_bytes() as i32) - before;`)
 	}
 	small := mustRunX86_64FreeOn(t, bumped("50"))
 	large := mustRunX86_64FreeOn(t, bumped("5000"))
