+++ internal/e2e/rc_string_recycle_test.go
@@ -29,7 +29,7 @@ func TestX86_64StringLocalRecycles(t *testing.T) {
         if (s.len() == 0) { return 99; }
         i = i + 1;
     }
-    if (__heap_bump_bytes() < 1048576) { return 0; }
+    if ((__heap_bump_bytes() as i32) < 1048576) { return 0; }
     return 1;
 }
 function main(): i32 { return churn("prefix", 100000); }`
@@ -43,7 +43,7 @@ function main(): i32 { return churn("prefix", 100000); }`
     var i: i32 = 0; var acc: i32 = 0;
     while (i < n) { acc = acc + (a + b + a + b).len(); i = i + 1; }
     if (acc == 0) { return 99; }
-    if (__heap_bump_bytes() < 1048576) { return 0; }
+    if ((__heap_bump_bytes() as i32) < 1048576) { return 0; }
     return 1;
 }
 function main(): i32 { return churn("longer_string_one_here", "longer_string_two_here", 100000); }`
