+++ internal/e2e/rc_threaded_array_param_test.go
@@ -44,7 +44,7 @@ func threadedArrayParamSrc(n, twoN string) string {
     return buf;
 }
 function main(): i32 {
-    var before: i32 = __heap_bump_bytes();
+    var before: i32 = (__heap_bump_bytes() as i32);
     var acc: i32[] = [];
     var i: i32 = 0;
     while (i < ` + n + `) { acc = le32(acc, i); i = i + 1; }
@@ -52,7 +52,7 @@ function main(): i32 {
     if (acc[0] != 0 || acc[2] != 1 || acc[6] != 3 || acc[7] != 0) { return 253; }
     if (__rc_underflow_count() != 0) { return 200; }
     if (__arr_push_shared_count() != 0) { return 201; }
-    return (__heap_bump_bytes() - before) / (acc.len() * 4);
+    return ((__heap_bump_bytes() as i32) - before) / (acc.len() * 4);
 }`
 }
