+++ internal/e2eselfhost/self_host_arrarr_reclaim_wasm_ir_test.go
@@ -47,14 +47,14 @@ func TestSelfHostArrArrReclaimWasmIR(t *testing.T) {
         acc = acc + g.len() + g[0][0].len();
         i = i + 1;
     }
-    var b1: i32 = __heap_bump_bytes();
+    var b1: i32 = (__heap_bump_bytes() as i32);
     var j: i32 = 0;
     while (j < 1500) {
         var g2: string[][] = [["a" + "b"], ["c" + "d", "e" + "f"]];
         acc = acc + g2.len() + g2[1][1].len();
         j = j + 1;
     }
-    var b2: i32 = __heap_bump_bytes();
+    var b2: i32 = (__heap_bump_bytes() as i32);
     if (__rc_underflow() != 0) { return 99; }
     if (b2 - b1 >= 4096) { return 98; }
     if (acc < 0) { return 97; }
@@ -68,14 +68,14 @@ func TestSelfHostArrArrReclaimWasmIR(t *testing.T) {
         acc = acc + g.len() + g[0][0];
         i = i + 1;
     }
-    var b1: i32 = __heap_bump_bytes();
+    var b1: i32 = (__heap_bump_bytes() as i32);
     var j: i32 = 0;
     while (j < 1500) {
         var g2: i32[][] = [[j, j + 1], [j + 2]];
         acc = acc + g2.len() + g2[1][0];
         j = j + 1;
     }
-    var b2: i32 = __heap_bump_bytes();
+    var b2: i32 = (__heap_bump_bytes() as i32);
     if (__rc_underflow() != 0) { return 99; }
     if (b2 - b1 >= 4096) { return 98; }
     if (acc < 0) { return 97; }
